Separated game creation and menu functionalities into modules
